requirements.txt
aws-cdk-lib>=2.248.0,<3.0.0
constructs>=10.5.0,<11.0.0
aws-lambda-powertools